requirements.txt
Flask>=2.0
numpy>=1.23
pillow>=9.0
tensorflow>=2.12,<2.22
flask-cors>=4.0
gunicorn>=21.0
matplotlib>=3.5
seaborn>=0.12
scikit-learn>=1.0
pathlib2>=2.3
